+++ src/scalecast/auxmodels.py
@@ -1,7 +1,101 @@
 from sklearn.neural_network import MLPRegressor
 from sklearn.ensemble import BaggingRegressor
 from sklearn.ensemble import StackingRegressor
+from statsmodels.tsa.vector_ar.vecm import VECM
 from scalecast.Forecaster import _sklearn_imports_
+
+class vecm:
+    def __init__(
+        self,
+        exog_coint=None,
+        k_ar_diff=1, # always 0
+        coint_rank=1,
+        deterministic="n",
+        seasons=0,
+        first_season=0,
+    ):
+        """ initializes a Vector Error Correction Model.
+        uses the statsmodels implementation: https://www.statsmodels.org/dev/generated/statsmodels.tsa.vector_ar.vecm.VECM.html
+
+        Args:
+            exog_coint (ndarray): default None. deterministic terms inside the cointegration relation.
+            k_ar_diff (int): default 1. number of lagged differences in the model.
+            coint_rank (int): cointegration rank.
+            deterministic (str): one of {"n", "co", "ci", "lo", "li"}. default "n".
+                "n" - no deterministic terms.
+                "co" - constant outside the cointegration relation.
+                "ci" - constant within the cointegration relation.
+                "lo" - linear trend outside the cointegration relation.
+                "li" - linear trend within the cointegration relation.
+                Combinations of these are possible (e.g. "cili" or "colo" for linear trend with intercept). 
+                When using a constant term you have to choose whether you want to restrict it to the cointegration relation 
+                (i.e. "ci") or leave it unrestricted (i.e. "co"). Do not use both "ci" and "co". The same applies for "li" 
+                and "lo" when using a linear term. 
+            seasons (int): default 0. Number of periods in a seasonal cycle. 0 means no seasons.
+            first_season (int): default 0. Season of the first observation.
+
+        Returns:
+            an initialized model.
+        """
+        self.k_ar_diff = k_ar_diff
+        self.exog_coint = exog_coint
+        self.coint_rank = coint_rank
+        self.deterministic = deterministic
+        self.seasons = seasons
+        self.first_season = first_season
+        self._scalecast_set = ['dates','freq','n_series'] # these attrs are set when imported into scalecast
+
+    def fit(self,X,y=None):
+        """ fits the model.
+
+        Args:
+            X (ndarray): the known observations (all known series plus exogenous regressors in a matrix).
+                MVForecaster will split endog from exog appropriately.
+            y: not required for this model and left None. kept in to be consistent
+                with other scikit-learn models syntax.
+        """
+        exog = X[:,self.n_series:]
+        X = X[:,:self.n_series]
+
+        self.exogdim = exog.shape[1]
+        self.Xdim = X.shape[1]
+
+        exog = None if exog.shape[1] == 0 else exog
+
+        self.mod = VECM(
+            X,
+            exog=exog,
+            k_ar_diff=self.k_ar_diff,
+            exog_coint=self.exog_coint,
+            coint_rank=self.coint_rank,
+            deterministic=self.deterministic,
+            seasons=self.seasons,
+            first_season=self.first_season,
+            dates = self.dates[:X.shape[0]],
+            freq = self.freq,
+        ).fit()
+        
+    def predict(self,X):
+        """ forecasts into an unknown horizon
+
+        Args:
+            X (ndarray): the future sereis values and future exognenous regressors.
+                MVForecaster will split this up appropriately without leaking.
+        """
+        # testing
+        if X.shape[1] == self.Xdim + self.exogdim:
+            exog = X[:,self.n_series:]
+            exog = None if exog.shape[1] == 0 else exog
+        # true forecast
+        elif X.shape[1] == self.exogdim:
+            exog = None if X.shape[1] == 0 else X
+        else: # i don't see how this would ever happen but i'd like to know if it does
+            raise Exception(
+                'something went wrong. ' 
+                'please consider raising this issue on the scalecast issues tab: '
+                'https://github.com/mikekeith52/scalecast/issues'
+            )
+        return self.mod.predict(steps=X.shape[0],exog_fc=exog)
 
 def auto_arima(f,call_me='auto_arima',Xvars=None,**kwargs):
     """ adds a forecast to a `Forecaster` object using the auto_arima function from pmdarima.